requirements.txt
fastapi==0.115.5
swagger-ui-bundle==1.1.0
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
bcrypt==4.2.1
PyJWT==2.10.1
python-multipart==0.0.20
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
